Debug static size hashtable, add extra warning flags (with -Werror.) Add more explicit conversions.

+++ src/include/graph.h
@@ -51,8 +51,8 @@ int       ndl_graph_set(ndl_graph *graph, ndl_ref node, ndl_sym key, ndl_value v
 ndl_value ndl_graph_get(ndl_graph *graph, ndl_ref node, ndl_sym key);
 int       ndl_graph_del(ndl_graph *graph, ndl_ref node, ndl_sym key);
 
-int     ndl_graph_size (ndl_graph *graph, ndl_ref node);
-ndl_sym ndl_graph_index(ndl_graph *graph, ndl_ref node, int index);
+int64_t ndl_graph_size (ndl_graph *graph, ndl_ref node);
+ndl_sym ndl_graph_index(ndl_graph *graph, ndl_ref node, int64_t index);
 
 
 /* Search backreferences.
@@ -62,8 +62,8 @@ ndl_sym ndl_graph_index(ndl_graph *graph, ndl_ref node, int index);
  * size() get the number of backreferences (including self.)
  * index() get the nth backreference.
  */
-int     ndl_graph_backref_size (ndl_graph *graph, ndl_ref node);
-ndl_ref ndl_graph_backref_index(ndl_graph *graph, ndl_ref node, int index);
+int64_t ndl_graph_backref_size (ndl_graph *graph, ndl_ref node);
+ndl_ref ndl_graph_backref_index(ndl_graph *graph, ndl_ref node, int64_t index);
 
 /* Print the entirety of a graph. */
 void ndl_graph_print(ndl_graph *graph);
